Refactor: 사용자 예약 로직 DTO 역할별 분리
