+++ src/main/java/com/acorn/api/controller/reservation/ReservationPageController.java
@@ -1,7 +1,8 @@
 package com.acorn.api.controller.reservation;
 
-import com.acorn.api.dto.reservation.ReservationDetailDTO;
-import com.acorn.api.dto.reservation.ReservationListDTO;
+import com.acorn.api.dto.reservation.response.ReservationDetailResDTO;
+import com.acorn.api.dto.reservation.request.ReservationListReqDTO;
+import com.acorn.api.dto.reservation.response.ReservationListResDTO;
 import com.acorn.api.service.reservation.ReservationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -18,8 +19,8 @@ public class ReservationPageController {
     private final ReservationService reservationService;
 
     @GetMapping("/user/reservation/list")
-    public String reservationListPage(ReservationListDTO listData, Model model) {
-        List<ReservationListDTO> reservationListData = reservationService.getReservationListData(listData);
+    public String reservationListPage(ReservationListReqDTO listData, Model model) {
+        List<ReservationListResDTO> reservationListData = reservationService.getReservationListData(listData);
         model.addAttribute("reservationListData", reservationListData);
         model.addAttribute("request", listData);
         return "reservation/reservation-list";
@@ -27,7 +28,7 @@ public class ReservationPageController {
 
     @GetMapping("/user/reservation/detail/{reservationId}")
     public String reservationDetailPage(@PathVariable Integer reservationId, Model model) {
-        ReservationDetailDTO reservationDetailData = reservationService.getReservationData(reservationId);
+        ReservationDetailResDTO reservationDetailData = reservationService.getReservationData(reservationId);
         model.addAttribute("reservationDetailData", reservationDetailData);
         return "reservation/reservation-detail";
     }
