+++ src/main/java/com/acorn/api/service/reservation/ReservationService.java
@@ -1,8 +1,9 @@
 package com.acorn.api.service.reservation;
 
-import com.acorn.api.dto.reservation.ReservationCancelDTO;
-import com.acorn.api.dto.reservation.ReservationDetailDTO;
-import com.acorn.api.dto.reservation.ReservationListDTO;
+import com.acorn.api.dto.reservation.request.ReservationCancelReqDTO;
+import com.acorn.api.dto.reservation.response.ReservationDetailResDTO;
+import com.acorn.api.dto.reservation.request.ReservationListReqDTO;
+import com.acorn.api.dto.reservation.response.ReservationListResDTO;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,11 +11,11 @@ import java.util.List;
 @Service
 public interface ReservationService {
 
-    List<ReservationListDTO> getReservationListData(ReservationListDTO listData);
+    List<ReservationListResDTO> getReservationListData(ReservationListReqDTO listData);
 
-    ReservationDetailDTO getReservationData(Integer reservationId);
+    ReservationDetailResDTO getReservationData(Integer reservationId);
 
     void reserveContainer(Integer containerId);
 
-    void reserveCancelContainer(ReservationCancelDTO requestData);
+    void reserveCancelContainer(ReservationCancelReqDTO requestData);
 }
